move back to server due to json infinite loop problem
