+++ server/src/main/java/org/schlunzis/kurtama/server/game/TerrainFactory.java
@@ -1,6 +1,6 @@
 package org.schlunzis.kurtama.server.game;
 
-import org.schlunzis.kurtama.common.game.model.ITerrain;
+import org.schlunzis.kurtama.server.game.model.ITerrain;
 
 public interface TerrainFactory {
