+++ server/src/main/java/org/schlunzis/kurtama/server/game/model/IDirection.java
@@ -1,4 +1,4 @@
-package org.schlunzis.kurtama.common.game.model;
+package org.schlunzis.kurtama.server.game.model;
 
 public interface IDirection {
